+++ src/Utils.h
@@ -8,6 +8,13 @@ namespace RedFS {
 
 using SharedMutex = std::shared_ptr<std::mutex>;
 
+static bool equals_insensitive(const std::string_view& p_a,
+                               const std::string_view& p_b) {
+  return std::ranges::equal(p_a, p_b, [](char a, char b) -> bool {
+    return std::tolower(a) == std::tolower(b);
+  });
 }
+
+}  // namespace RedFS
 
 #endif  //REDFILESYSTEM_UTILS_H